task_manager/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-ratelimit
celery
redis
django-celery-beat